msm: camera: tpg: Updation of ASYNC_MODE_MIN_HBI param under tpg_ctrl
TPG HW on Canoe has a HW bug that requires this
register field to be written with an hardcoded value
as suggested by the hardware team. The corresponding
bitfields in that register are not used on Lanai and
hence there is no need to make this change target specific.
The change include updating bitfield 23:16 with 0xA value of
TPG_CTRL Register.

CRs-Fixed: 3827853
External Impact: No
Change-Id: Ia807b25fc200be85b2c6ad29c9ce333bc0210028
Signed-off-by: Abhilash Mahapatra <quic_abhmah@quicinc.com>