disp: msm: dp: check for any lttprs in the PHY layer and set the mode
In the PHY layer of a DP connection (cable or/and the sink), often
"Link Training(LT) tunable PHY repeaters (LTTPR)" are employed.
These LTTPRs can operate in 2 modes: Transparent & Non-Transparent.
Even though the DP 1.4spec suggests transparent mode as default for
LTTPRs, it is observed that some cables with LTTPRs misbehave if the
operating mode isn't set explicitly by the DP TX. This change set the
transparent mode if at least 1 LTTPR is present in the PHY layer.

Change-Id: I334b1779df6be47616cd48ae7f33df8af4af4036
Signed-off-by: Sandeep Gangadharaiah <quic_sandgang@quicinc.com>