disp: msm: sde: enable cesta auto-active-on-panic for all frames
In command mode, when the previous frame update ctl-done is very
close to panic/wakeup windows, there is a possible race condition
leading to idle-vote. As a workaround, enable auto-active-on-panic
before every frame-update along with force-db-updated for it to take
effect. This will make sure, idle vote is not applied and remove the
vote after the wr-ptr-irq as part of the complete-commit.

Change-Id: I6db924c9cc521ca8a86b060de5c087d2bb4e7fc7
Signed-off-by: Veera Sundaram Sankaran <quic_veeras@quicinc.com>