disp: msm: sde: enable cesta auto-active-on-panic for all frames
In command mode, when the previous frame update ctl-done is very
close to panic/wakeup windows, there is a possible race condition
leading to idle-vote. As a workaround, enable auto-active-on-panic
before every frame-update along with force-db-updated for it to take
effect. This will make sure, idle vote is not applied and remove the
vote after the wr-ptr-irq as part of the complete-commit.

Signed-off-by: Veera Sundaram Sankaran <quic_veeras@quicinc.com>
Change-Id: I9210839c7537ec67d4bbfa991473c4e3541a5cdc
Signed-off-by: Xhoendi Collaku <xcollaku@quicinc.com>
Signed-off-by: Xhoendi Collaku <quic_xcollaku@quicinc.com>