disp: msm: sde: ensure MDSS goes to PC when idle
When userspace requests a vsync interrupt timestamp, driver adds a
power vote and holds it until it returns the timestamp. This change
ensures that if we're already in the middle of the power collapse
procedure, CRTC will wait for it to complete, thus clearing the
hardware registers, before waking it up for the vsync timestamp.

Change-Id: I480be0c6c4fe42e0a238ec3f0d538a648ad0bb04
Signed-off-by: Kirill Shpin <quic_kshpin@quicinc.com>