mm-drivers: hw_fence: always remove power votes on soccp for ctrlq msg
SoCCP sends ctrl rxq messages asynchronously back to HLOS, so HW Fence
Driver should not rely on SoCCP to send this message back before
removing power vote. Instead, wait for this message with timeout and
remove the power vote regardless if message is received or not received
by this timeout.

Change-Id: I37cfeefba0f799949609fc30af0d0bff9c30c3d0
Signed-off-by: Grace An <quic_gracan@quicinc.com>