msm: camera: isp: Add bounds check for irq reg val
KASAN: slab out of bounds in cam_vfe_bus_ver3_out_done_top_half_util
Kasan tests a scenario where the num of registers is higher than
possible which is not considered in the design of this code. A check
is added as a form of defensive coding which should take care of this
issue.

CRs-Fixed: 3819301
Change-Id: Ib48b42e4f481d81f318e1f4353490454db2821be
Signed-off-by: Atiya Kailany <quic_akailany@quicinc.com>