disp: msm: add cesta mode2_entry status check
As part of the power-collapse sequence, cesta triggers the mode-2
sequence and GDSC power-collapse when the last sw-client vote is
removed. Add a polling logic to check the GDSC power-collapse
status after the last sw-client vote removal to confirm all cesta
activity is finished before requesting for MMCX power collapse.

Change-Id: I0097e625d01c97b167816b96e42b14a0eb981a0b
Signed-off-by: Veera Sundaram Sankaran <quic_veeras@quicinc.com>