disp: msm: sde: add flag for dsi_prop panel events
Display driver currently sends a notification to other drivers
such as touch for events that may concern them, like suspend.
The touch driver's list of panels to activate upon may include the
default panel, which will be formally picked by display driver, but
in practice the data in the device tree would be substituted with
the dsi_prop file, resulting in a mismatch between these drivers.

Add device tree flag, to be used exclusively in dsi_prop files, to
disable panel event notifications to other drivers.

Change-Id: I13c191e65fb574569920f174caeb144c9a22209d
Signed-off-by: Kirill Shpin <quic_kshpin@quicinc.com>