msm: camera: cpas: Fix bw consolidation for drv based bw voting
Currently, during consolidation of bw, the vote level is picked
based on first entry in the incoming client vote. This assumption
is incorrect, as different paths combining to a consolidated path
can have different vote levels. Fix this by adding a separate
structure for consolidate vote and change the corresponding logic
to accommodate both high and low levels accordingly.

CRs-Fixed: 3850663
Change-Id: I70ec77f05982cfe6974b47d73af7cdfc1943f6ef
Signed-off-by: Mukund Madhusudan Atre <quic_matre@quicinc.com>