disp: msm: sde: modify cesta sequence for cmd mode fps switch
On FPS switch frames modify the cesta voting req to immediate
from default panic_region. This will make sure the voting takes
effect immediately on these cases. Currently, on FPS switch cases
from lower to higher fps, the tear-check vsync-count which is a
single buffer register is updated soon and this make the cesta
panic/wakeup windows to fire at the intended lines, but later in
the timeline. This leads to very little time for clk/bw update
after the panic and leads to continuously dropping the frame.

Change-Id: I92c3434423b4bed63fcd29a7e1eeb34238ef8ce6
Signed-off-by: Veera Sundaram Sankaran <quic_veeras@quicinc.com>