disp: msm: dsi: move esync clk parsing
Esync clock's parent needs to be changed to the DSI PLL exactly
once, which is best done when initializing all the clocks, and
in a location where both the esync clock RCG and the to-be-parent
clock (DSI PLL) are accessible. This must also now happen after
parsing panel device tree information, such that we know whether
esync will be necessary for this panel.

Move the esync clock parsing and init from dsi_ctrl to dsi_display,
by which point panel information will have already been parsed.

Change-Id: Id2f58a0bbe62f87b1b7ef6b4e5172f5f60c82593
Signed-off-by: Kirill Shpin <quic_kshpin@quicinc.com>