disp: msm: dsi: move esync clk parsing
Esync clock's parent needs to be changed to the DSI PLL exactly
once, which is best done when initializing all the clocks, and
in a location where both the esync clock RCG and the to-be-parent
clock (DSI PLL) are accessible. This must also now happen after
parsing panel device tree information, such that we know whether
esync will be necessary for this panel.

Move the esync clock parsing and init from dsi_ctrl to dsi_display,
by which point panel information will have already been parsed.

Change-Id: I1faa236f8418550f8aa0eb738e6ef154ae73c9f8
Signed-off-by: Kirill Shpin <quic_kshpin@quicinc.com>