disp: msm: sde: force hard reset when cesta is enabled
Currently, ctl soft-reset is done first as part of the hw reset
sequence for recovering DPU from various hang/stuck issues. If
this step is successful, it returns early. This is not sufficient
for cesta SCC to recover. Force partial hard-reset sequence
all times when cesta is enabled to reset the cesta SCC cleanly.

Change-Id: I9a1a5fa6b83710cdba4d61a0cf882155381097ce
Signed-off-by: Veera Sundaram Sankaran <quic_veeras@quicinc.com>