dsp-kernel: Retrieve the map size from the map table
The current code relies on the map size from the user passed length,
which is untrusted for validating the input buffer length and could
potentially lead to buffer out of bounds access. The fix is to obtain
the map size from the map table instead of depending on the user passed
length.

Change-Id: I657b1ced294d19415e6580dc3050eeffff464324
Signed-off-by: Ramesh Nallagopu <quic_rnallago@quicinc.com>