dsp-kernel: Add support for waking up notification thread
When a user makes an ioctl call to wake up the notification thread in
the driver, it fails because the kernel does not support this wakeup.
Due to this failure, userspace attempts to force kill the notification
thread, causing a race condition and making fastrpc_apps_user_deinit
get stuck. Adding support for waking up the notification thread in
the driver

Change-Id: I528520fc26d404131d6d590999e5b07af9bc7eb8
Signed-off-by: Ramesh Nallagopu <quic_rnallago@quicinc.com>