msm: camera: reqmgr: Remove the initial frame skipping
The initial frame skipping was added to fix the issue if one of the
pipeline stream on too early on some old target,but we don't have
such issue now and it always drop one frame during start up which
affects the start up performance. So, this change removes this
initial frame skipping to improve the performance
And also keep this checking if own current request is faster than
peer link by more than one.

CRs-Fixed: 3878262
Change-Id: I6187541772f0a25a2c85a1ada9842948e0d8fb4a
Signed-off-by: Depeng Shao <quic_depengs@quicinc.com>