dsp-kernel: Remove unnecessary kfree for notif
In the fastrpc_wait_on_notif_queue function, if any error occurs
during the wait for the notif event, kfree is called for the notif
pointer, which is not yet initialized. This kfree is not required
since notif is not yet assigned and there is nothing to be freed.
In an error scenario just return the error condition. Remove kfree
for notif and initialize notif with null. This ensures that the check
for a valid notif is not affected by a garbage value.

Change-Id: I30335e8d1ccaf6d287190819d4f01ab9a7072976
Signed-off-by: ANANDU KRISHNAN E <quic_anane@quicinc.com>