qcacld-3.0: Wait for vdev restart complete in set_key command
Currently host driver rejects set_key if vdev_restart is in
progress which eventually triggers stop_ap from userspace leading
to SAP turn on failure.
In current issue, host driver receives unsafe channel event during
SAP start and the frequency on which SAP is getting started is also
part of the unsafe channel event. Once SAP start completes, host
driver triggers CSA as the current channel is unsafe. As part of
CSA, vdev restart gets triggers and when restart is in progress,
userspace issues set_key command and host rejects this command.
As a rejection of set_key command, userspace triggers stop_ap
which results in SAP turn on failure.

To fix this issue, wait for vdev restart complete in set_key if
there is any CSA is in progress.

Change-Id: Id1d4fd38754b4b443d62b2976385b083ba3833ff
CRs-Fixed: 3895619