disp: msm: sde: wait for prev frame before sec-cam preview session
Secure camera preview and secure playback usecases share the same SID
used in a different way by repurposing it from nested to S2-only through
scm_call. As part of the switch, all the secure-buffers are unmapped.
Before the switch, make sure to wait for the previous frame
transfer to complete to avoid any early unmapping.

Change-Id: I05934f48228bc1710309e33f5cd5a1a4fcfbbbf4
Signed-off-by: Veera Sundaram Sankaran <quic_veeras@quicinc.com>