msm: camera: core: Add check to match context with node in acquire hw
Currently, we just check the handle table to see if the session
handles and device handles match, but this does not check for
cases where we give a valid device handle, but for a non-matching
device (node). This change adds this check to prevent potential
incorrect memory casting and reference.

CRs-Fixed: 3886586
Change-Id: Ief59831a52f6a6928e3159003a38477c8843e1e9
Signed-off-by: Li Sha Lim <quic_lishlim@quicinc.com>