mm-drivers: hw_fence: propagate error from failed rx queue update
If Rx queue is full at time that client registers for wait on an
already signaled fence, the rx queue will fail to be updated. Currently,
this error is ignored. Instead, propagate this error to the client and
also remove the client's refcount in this scenario so that the client
consistently does not need to invoke synx_release after synx_import
call fails.

Change-Id: Ic723d20ac5d1a769a5afe3b6ad7efab5191cbc5e
Signed-off-by: Grace An <quic_gracan@quicinc.com>