dsp-kernel: Handle race-condition in dsp signal
The `fastrpc_dspsignal_wait` function currently checks the
signal state before waiting for a signal from the DSP. However,
if the signal is already received before the check, it results
in an infinite loop, causing excessive resource usage.

This change addresses the race condition by checking both the
pending and signaled states. If the signal is not in the pending
state, it directly checks for the signaled state, resets the states,
and returns to avoid looping.

Change-Id: I7aaf4e8f2150b6899cf9788667bba002d18b2e4d
Signed-off-by: ANANDU KRISHNAN E <quic_anane@quicinc.com>