qcacld-3.0: Cache the USD adapter
Driver can receives USD vendor command on STA or P2P interface from
the userspace and driver need to send the USD multicast follow up
frames to the interface on which it received the USD vendor command.

So, to identify the interface to send follow up frames, caches
adapter in HDD context, for which driver receives the USD vendor
command, to use it later.

Change-Id: If4d50d889b84e861672464569374d28598408f19
CRs-Fixed: 3910953