qcacmn: Change logging level for excess logging in kmsg
In reg_chan_opclass_to_freq_auto function, gets the frequency
with op_class firstly from the current country table if not
present then looks up in global table. If the op_class not
present in current country table and present in global table,
gives the error log for the first case, but anyway able to find
the same op_class in the global table.

This will flood the logs in success case also. To solve this
remove error log in calling function and add error log with debug
level in the caller functions.

Change-Id: I8980f9fc068069c726284522ef1ef6944075eb18
CRs-Fixed: 3903815