qcacld-3.0: Add check for the frequency in callers
In reg_chan_opclass_to_freq_auto function, gets the frequency
with op_class firstly from the current country table if not
present then looks up in global table. If the op_class not
present in current country table and present in global table,
gives the error log for the first case, but anyway able to find
the same op_class in the global table.

This will flood the logs in success case also. To solve this
remove error log in calling function and add error log with debug
level in the caller functions.

Change-Id: Ia3f3b7ffba19f2170fbb152d02a955e8f6505c57
CRs-Fixed: 3914337