disp: msm: sde: add smmu_proxy_switch_sid call before scm_call
As part of CSF 2.5 usecase end frame, add a call to smmu_proxy_switch_sid
to release the SIDs from TVM. Previously, SID release was done as part
of the last buffer free. Freeing the SIDs early allows the PVM to use
the SID early and not be dependent on the app/sf to free the buffer.
This helps in secure-camera preview followed by secure playback,
back-to-back usecase.

Change-Id: I10cda2d0f6ca0fbe969180b5396c9f4b5aa818bb
Signed-off-by: Veera Sundaram Sankaran <quic_veeras@quicinc.com>