qcacld-3.0: Disable near full irq feature for orne
Currently, if HW is going into low power mode and few
Rx packets are present in ring, only MSI2 (near full)
interrupt is raised which is not used to process the Rx
ring. This is causing high latency as the packets are not
getting processed until HW is out of low power mode and a
REO MSI interrupt is raised.
To fix this, disable CONFIG_WLAN_FEATURE_NEAR_FULL_IRQ for Orne.

CRs-Fixed: 3918846
Change-Id: I83195060d2a66c1a7b55090c4800bd7c69cb4661