qcacld-3.0: Validate established link_id if establish TID mapping is present
Currently, during candidate validation, without verifying the
established TID mapping value, the driver matches the
established_tid mapped_link_id with candidate self_link_id.
If they match, it proceeds otherwise, it rejects the candidate.

It may be possible that the TTLM mapping switch timer is
running and the TID mapping is present in the upcoming TID mapping,
not in the established TID mapping.
In this case, the established TID mapping is 0, and while converting
TID mapping to LINK_ID, the driver will get link_id: 0
(which might be pointing to the 2.4 GHz link_id), which is incorrect.
This causes the driver to reject the current candidate and select
the next candidate for connection.

Fix is, if the established TID mapping is 0, the driver should return
SUCCESS and proceed with the same candidate.

CRs-Fixed: 3921701
Change-Id: Id4fa1409bc7abafa4cdb231d5c33ab72f7b9b9ea