qcacld-3.0: Don't reject the twt setup for 2.4 GHz connection
Currently driver sends the TWT requestor capability to userspace
for MLO connection. For non-MLO connection, this caps is not advertised
if connection is in 2.4 GHz and enable_twt_24ghz ini is 0.
Now if TWT requestor caps is not advertised and userspace sends the
TWT setup command the driver rejects it.
This causes the cert issue where this TWT requestor cap is not checked
and TWT setup command is issued on 2.4 GHz connection.

So, to resolve this issue, don't reject the TWT setup command if
connection is on 2.4 GHz and enable_twt_24ghz is 0.

Change-Id: I5ebc3fcd0c28612358c6a810752f6c558fb23cd6
CRs-Fixed: 3910377