dsp-kernel: Fix unnecessary ownership reassignment of remoteheap
The Audio PD daemon allocates memory for audio PD runtime dynamic
loading usage. As part of this, the memory ownership moves to the
VM where audio PD can use it. In a memory ownership assignment error
scenario as a part of err_assign unnecessary ownership reassignment
happens, which fails and adds to the buffer in the gmaps list,
remaining unused and resulting in a memory issue. Avoid unnecessary
ownership reassignments of remoteheap.

Change-Id: Iae6fd27b3171e3cb8a04d5c1c65a050304ca4cd2
Signed-off-by: Santosh <quic_ssakore@quicinc.com>