qcacld-3.0: Fix for freq to chan mapping in set time quota
Currently for frequency to channel mapping in set time quota
command using the legacy API which support 2.4 GHz and 5 GHz
bands. For 6 GHz band returns the invalid channel which leads
to crash in FW.

Change the legacy API to regulatory API supports all the bands
and while posting message to scheduler add band mask for the two
frequencies to help in convert back from channel to frequency.

Change-Id: If4038125ffaab38d057b0d97444cd1269fde18db
CRs-Fixed: 3921964