qcacld-3.0: Don't disable ML links for ML-STA+NAN+P2P concurrency
When two legacy connections are present or second legacy
connection comes up on different channels but on same mac and
if an ML link is active on that mac, MLO mgr makes that link
inactive to avoid 3-channel MCC on a mac.
But NAN supports 3-port for non LL cases as below,
1. 3-port on a mac with all channel SCC
2. 3-port on a mac with 2-channel MCC(e.g. port-1 on 5180, port-2
   and port-3 on 5745)
3. 3-port on a mac with 3-channel MCC. Firmware moves the NDP to
   other mac to handle this. So, host can allow and firmware
   dynamically takes decision.

So, don't make the STA link inactive though it's leading to 3-port
when NAN and non-LL P2P are present.

Change-Id: I8c0509ed1d7c45eb37519b13b2b44093572079ce
CRs-Fixed: 3925417