dsp-kernel: Fix use after free buf in init process create
When two threads simultaneously make ioctl calls to create the process,
Thread1 creates and uses the buffer. Meanwhile, Thread2 can access and
free the same buffer in a failure case, causing Thread1 to still access
it, resulting in a use-after-free issue. To prevent this, use a file
state and reject the process creation if it is already in progress.

Change-Id: I33a3026451cf9f2989519659c937ed17224612ce
Signed-off-by: Ramesh Nallago <quic_rnallago@quicinc.com>