qcacmn: Fix smmu unmmaping of ctrl pkt
There is a race condition between
the SMMU mapping of control packets and the SMMU
pool unmapping in opt_dp.

control packets either residing in or en route
to the REO2SW4 ring become unmapped as part of
pool unmapping leading SMMU fault.

To resolve this issue ensuring that control packets
are not unmapped during the SMMU pool unmapping process
by adding flag for ctrl pkt in rx descriptor.

Change-Id: Ie55a419f46d157168a6d4f90e5e599fa1082aa1c
CRs-Fixed: 3935382