qcacld-3.0: Fix disabled link moved to connection table
In non-dbs capable hardware only one active link entry should be
present for ML STA. But when set link active response is
received from firmware it doesn't carry the inactive_link TLV if
requested mode is 5 (No_Force). So in that case the inactive
link bitmap is 0 in the
policy_mgr_handle_vdev_active_inactive_resp() API.
The active link bitmap is obtained as !inactive_linkid_bitmap,
so it causes all links to be active and disabled links are also
moved to connection table. So get the
curr_inactive_linkid_bitmap in case of the request mode is 5
(no_force).

Also after roaming move the inactive link to disabled table.

CRs-Fixed: 3937046
Change-Id: I62c12298a98f39f2ddc07b6ad6fd4505c544e72a